Update target framework and packability settings
The properties of 'TargetFramework' and 'IsPackable' were updated in Directory.Build.props, Frank.Templates.Generator.csproj, and Frank.Templates.Generator.Tests.csproj files. These changes have been made to ensure compatibility with the .NET 7.0 framework and to set packability of the projects as needed. The Generator project is now marked as packable, while the Tests project is not, aligning with common .NET Project configuration practices.