Update NugetSolution template metadata
The metadata for the NugetSolution template in the template.json file was updated. This includes changing the classifications, identity, name, description, and sourceName according to the purpose of the new library. Additionally, the type tag was updated to correctly reflect that the template is for a solution, not a project.